+++ packages/brick_graphql/CHANGELOG.md
@@ -1,5 +1,13 @@
 ## Unreleased
 
+## 0.0.1+2
+
+* Rename `fieldsToRuntimeDefinition` to `fieldsToGraphqlRuntimeDefinition`
+
+## 0.0.1+1
+
+* Return `List<_Model>` when invoking `subscribe`
+
 ## 0.0.1
 
 Alpha release