+++ packages/brick_offline_first_with_graphql/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+* Add `subscribe` for streaming updates of all models
+
 ## 0.0.1
 
 Initial